refactor(notebooks): extract the helper-cell template
notebook_export_support.py was 3592 lines, and `_helper_cell` was 1258 of
them: one f-string holding the Python that exported notebooks run to
locate their app, rebuild stage context, and render analysis pages.

It shares nothing with the module around it. The only value interpolated
from the caller is `payload_literal`; all 72 other brace pairs are `{{`
escaped and belong to the generated code. Static analysis of the function
body finds no free name defined in the module -- it needs json, textwrap
and Any, nothing else.

Move it verbatim to notebook_helper_cell.py. The string is untouched, so
exported notebooks are unchanged: rendering three payloads, including one
with unicode and literal braces, gives byte-identical output to the
pre-split function (sha256 compared against the version in git).

notebook_export_support.py drops 3592 -> 2334, below even the 3100 ceiling
that predated #807, so the budget comes down 3700 -> 2450 in the same
change. A ratchet that only ever loosens is not a ratchet; paying the debt
should reclaim the headroom rather than leave it to be refilled quietly.

notebook_helper_cell.py is left unbudgeted on purpose -- it is emitted
source for exported notebooks, not control flow anyone reads top to bottom.
